backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.30
